+++ src/Symfony/Bundle/FrameworkBundle/Tests/Console/ApplicationTest.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Tests\Console;
 
+use PHPUnit\Framework\MockObject\MockObject;
+use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Bundle\FrameworkBundle\Tests\TestCase;
 use Symfony\Component\Console\Command\Command;
@@ -20,39 +22,83 @@ use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\ApplicationTester;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+use Symfony\Component\HttpKernel\Bundle\BundleInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
 
 class ApplicationTest extends TestCase
 {
     public function testBundleInterfaceImplementation()
     {
-        $bundle = $this->getMockBuilder('Symfony\Component\HttpKernel\Bundle\BundleInterface')->getMock();
+        $bundle = $this->getMockBuilder(BundleInterface::class)->getMock();
 
-        $kernel = $this->getKernel(array($bundle), true);
+        $kernel = $this->getKernel([$bundle], true);
 
         $application = new Application($kernel);
-        $application->doRun(new ArrayInput(array('list')), new NullOutput());
+        $application->doRun(new ArrayInput(['list']), new NullOutput());
+    }
+
+    public function testApplicationIsAvailableAsService(): void
+    {
+        $bundle = $this->createBundleMock([]);
+        $kernel = $this->getKernel([$bundle], true);
+
+        /** @var ContainerInterface|MockObject $container */
+        $application = new Application($kernel);
+        $container = $kernel->getContainer();
+        $container->expects(self::once())
+            ->method('set')
+            ->with('application', $application);
+
+        $container->expects(self::once())
+            ->method('initialized')
+            ->with('application')
+            ->willReturn(false);
+
+        $application->doRun(new ArrayInput(['list']), new NullOutput());
+    }
+
+    public function testApplicationIsNotRegisteredTwice(): void
+    {
+        $bundle = $this->createBundleMock([]);
+        $kernel = $this->getKernel([$bundle], true);
+
+        /** @var ContainerInterface|MockObject $container */
+        $application = new Application($kernel);
+        $container = $kernel->getContainer();
+        $container->expects(self::never())
+            ->method('set')
+            ->with('application', $application);
+
+        $container->expects(self::once())
+            ->method('initialized')
+            ->with('application')
+            ->willReturn(true);
+
+        $application->doRun(new ArrayInput(['list']), new NullOutput());
     }
 
     public function testBundleCommandsAreRegistered()
     {
-        $bundle = $this->createBundleMock(array());
+        $bundle = $this->createBundleMock([]);
 
-        $kernel = $this->getKernel(array($bundle), true);
+        $kernel = $this->getKernel([$bundle], true);
 
         $application = new Application($kernel);
-        $application->doRun(new ArrayInput(array('list')), new NullOutput());
+        $application->doRun(new ArrayInput(['list']), new NullOutput());
 
         // Calling twice: registration should only be done once.
-        $application->doRun(new ArrayInput(array('list')), new NullOutput());
+        $application->doRun(new ArrayInput(['list']), new NullOutput());
     }
 
     public function testBundleCommandsAreRetrievable()
     {
-        $bundle = $this->createBundleMock(array());
+        $bundle = $this->createBundleMock([]);
 
-        $kernel = $this->getKernel(array($bundle));
+        $kernel = $this->getKernel([$bundle]);
 
         $application = new Application($kernel);
         $application->all();
@@ -65,9 +111,9 @@ class ApplicationTest extends TestCase
     {
         $command = new Command('example');
 
-        $bundle = $this->createBundleMock(array($command));
+        $bundle = $this->createBundleMock([$command]);
 
-        $kernel = $this->getKernel(array($bundle));
+        $kernel = $this->getKernel([$bundle]);
 
         $application = new Application($kernel);
 
@@ -78,9 +124,9 @@ class ApplicationTest extends TestCase
     {
         $command = new Command('example');
 
-        $bundle = $this->createBundleMock(array($command));
+        $bundle = $this->createBundleMock([$command]);
 
-        $kernel = $this->getKernel(array($bundle));
+        $kernel = $this->getKernel([$bundle]);
 
         $application = new Application($kernel);
 
@@ -90,11 +136,11 @@ class ApplicationTest extends TestCase
     public function testBundleCommandCanBeFoundByAlias()
     {
         $command = new Command('example');
-        $command->setAliases(array('alias'));
+        $command->setAliases(['alias']);
 
-        $bundle = $this->createBundleMock(array($command));
+        $bundle = $this->createBundleMock([$command]);
 
-        $kernel = $this->getKernel(array($bundle));
+        $kernel = $this->getKernel([$bundle]);
 
         $application = new Application($kernel);
 
@@ -107,30 +153,30 @@ class ApplicationTest extends TestCase
      */
     public function testBundleCommandsHaveRightContainer()
     {
-        $command = $this->getMockForAbstractClass('Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand', array('foo'), '', true, true, true, array('setContainer'));
+        $command = $this->getMockForAbstractClass(ContainerAwareCommand::class, ['foo'], '', true, true, true, ['setContainer']);
         $command->setCode(function () {});
         $command->expects($this->exactly(2))->method('setContainer');
 
-        $application = new Application($this->getKernel(array(), true));
+        $application = new Application($this->getKernel([], true));
         $application->setAutoExit(false);
         $application->setCatchExceptions(false);
         $application->add($command);
         $tester = new ApplicationTester($application);
 
         // set container is called here
-        $tester->run(array('command' => 'foo'));
+        $tester->run(['command' => 'foo']);
 
         // as the container might have change between two runs, setContainer must called again
-        $tester->run(array('command' => 'foo'));
+        $tester->run(['command' => 'foo']);
     }
 
     public function testBundleCommandCanOverriddeAPreExistingCommandWithTheSameName()
     {
         $command = new Command('example');
 
-        $bundle = $this->createBundleMock(array($command));
+        $bundle = $this->createBundleMock([$command]);
 
-        $kernel = $this->getKernel(array($bundle));
+        $kernel = $this->getKernel([$bundle]);
 
         $application = new Application($kernel);
         $newCommand = new Command('example');
@@ -144,14 +190,14 @@ class ApplicationTest extends TestCase
         $container = new ContainerBuilder();
         $container->register('event_dispatcher', EventDispatcher::class);
         $container->register(ThrowingCommand::class, ThrowingCommand::class);
-        $container->setParameter('console.command.ids', array(ThrowingCommand::class => ThrowingCommand::class));
+        $container->setParameter('console.command.ids', [ThrowingCommand::class => ThrowingCommand::class]);
 
         $kernel = $this->getMockBuilder(KernelInterface::class)->getMock();
         $kernel
             ->method('getBundles')
-            ->willReturn(array($this->createBundleMock(
-                array((new Command('fine'))->setCode(function (InputInterface $input, OutputInterface $output) { $output->write('fine'); }))
-            )));
+            ->willReturn([$this->createBundleMock(
+                [(new Command('fine'))->setCode(function (InputInterface $input, OutputInterface $output) { $output->write('fine'); })]
+            )]);
         $kernel
             ->method('getContainer')
             ->willReturn($container);
@@ -160,7 +206,7 @@ class ApplicationTest extends TestCase
         $application->setAutoExit(false);
 
         $tester = new ApplicationTester($application);
-        $tester->run(array('command' => 'fine'));
+        $tester->run(['command' => 'fine']);
         $output = $tester->getDisplay();
 
         $this->assertSame(0, $tester->getStatusCode());
@@ -177,9 +223,9 @@ class ApplicationTest extends TestCase
         $kernel = $this->getMockBuilder(KernelInterface::class)->getMock();
         $kernel
             ->method('getBundles')
-            ->willReturn(array($this->createBundleMock(
-                array((new Command(null))->setCode(function (InputInterface $input, OutputInterface $output) { $output->write('fine'); }))
-            )));
+            ->willReturn([$this->createBundleMock(
+                [(new Command(null))->setCode(function (InputInterface $input, OutputInterface $output) { $output->write('fine'); })]
+            )]);
         $kernel
             ->method('getContainer')
             ->willReturn($container);
@@ -188,7 +234,7 @@ class ApplicationTest extends TestCase
         $application->setAutoExit(false);
 
         $tester = new ApplicationTester($application);
-        $tester->run(array('command' => 'fine'));
+        $tester->run(['command' => 'fine']);
         $output = $tester->getDisplay();
 
         $this->assertSame(1, $tester->getStatusCode());
@@ -201,15 +247,15 @@ class ApplicationTest extends TestCase
         $container = new ContainerBuilder();
         $container->register('event_dispatcher', EventDispatcher::class);
         $container->register(ThrowingCommand::class, ThrowingCommand::class);
-        $container->setParameter('console.command.ids', array(ThrowingCommand::class => ThrowingCommand::class));
+        $container->setParameter('console.command.ids', [ThrowingCommand::class => ThrowingCommand::class]);
 
         $kernel = $this->getMockBuilder(KernelInterface::class)->getMock();
         $kernel->expects($this->once())->method('boot');
         $kernel
             ->method('getBundles')
-            ->willReturn(array($this->createBundleMock(
-                array((new Command('fine'))->setCode(function (InputInterface $input, OutputInterface $output) { $output->write('fine'); }))
-            )));
+            ->willReturn([$this->createBundleMock(
+                [(new Command('fine'))->setCode(function (InputInterface $input, OutputInterface $output) { $output->write('fine'); })]
+            )]);
         $kernel
             ->method('getContainer')
             ->willReturn($container);
@@ -218,7 +264,7 @@ class ApplicationTest extends TestCase
         $application->setAutoExit(false);
 
         $tester = new ApplicationTester($application);
-        $tester->run(array('command' => 'list'));
+        $tester->run(['command' => 'list']);
 
         $this->assertSame(0, $tester->getStatusCode());
         $display = explode('Lists commands', $tester->getDisplay());
@@ -226,12 +272,12 @@ class ApplicationTest extends TestCase
         $this->assertContains(trim('[WARNING] Some commands could not be registered:'), trim($display[1]));
     }
 
-    private function getKernel(array $bundles, $useDispatcher = false)
+    private function getKernel(array $bundles, bool $useDispatcher = false): KernelInterface
     {
-        $container = $this->getMockBuilder('Symfony\Component\DependencyInjection\ContainerInterface')->getMock();
+        $container = $this->getMockBuilder(ContainerInterface::class)->getMock();
 
         if ($useDispatcher) {
-            $dispatcher = $this->getMockBuilder('Symfony\Component\EventDispatcher\EventDispatcherInterface')->getMock();
+            $dispatcher = $this->getMockBuilder(EventDispatcherInterface::class)->getMock();
             $dispatcher
                 ->expects($this->atLeastOnce())
                 ->method('dispatch')
@@ -246,17 +292,17 @@ class ApplicationTest extends TestCase
         $container
             ->expects($this->exactly(2))
             ->method('hasParameter')
-            ->withConsecutive(array('console.command.ids'), array('console.lazy_command.ids'))
+            ->withConsecutive(['console.command.ids'], ['console.lazy_command.ids'])
             ->willReturnOnConsecutiveCalls(true, true)
         ;
         $container
             ->expects($this->exactly(2))
             ->method('getParameter')
-            ->withConsecutive(array('console.lazy_command.ids'), array('console.command.ids'))
-            ->willReturnOnConsecutiveCalls(array(), array())
+            ->withConsecutive(['console.lazy_command.ids'], ['console.command.ids'])
+            ->willReturnOnConsecutiveCalls([], [])
         ;
 
-        $kernel = $this->getMockBuilder('Symfony\Component\HttpKernel\KernelInterface')->getMock();
+        $kernel = $this->getMockBuilder(KernelInterface::class)->getMock();
         $kernel->expects($this->once())->method('boot');
         $kernel
             ->expects($this->any())
@@ -274,7 +320,7 @@ class ApplicationTest extends TestCase
 
     private function createBundleMock(array $commands)
     {
-        $bundle = $this->getMockBuilder('Symfony\Component\HttpKernel\Bundle\Bundle')->getMock();
+        $bundle = $this->getMockBuilder(Bundle::class)->getMock();
         $bundle
             ->expects($this->once())
             ->method('registerCommands')